+++ src/locus_cli.erl
@@ -39,15 +39,15 @@
 main(Args) ->
     ensure_apps_are_started([locus, getopt]),
     case Args of
-        ["analyze-robustness" | CmdArgs] ->
-            handle_robustness_analysis_command(CmdArgs);
+        ["analyze" | CmdArgs] ->
+            handle_analysis_command(CmdArgs);
         _ ->
             fall_from_grace(
               "~n"
               "Usage: locus [<command>] [<command_args>]~n"
               "~n"
               "Available commands:~n"
-              "  analyze-robustness")
+              "  analyze")
     end.
 
 %% ------------------------------------------------------------------
@@ -61,7 +61,7 @@ ensure_apps_are_started(Apps) ->
       end,
       Apps).
 
-handle_robustness_analysis_command(CmdArgs) ->
+handle_analysis_command(CmdArgs) ->
     OptSpecList =
         [{load_timeout, undefined, "load-timeout", {integer,30}, "Database load timeout (in seconds)"},
          {log_level,    undefined, "log-level",    {string,"error"}, "debug | info | warning | error"},
@@ -75,14 +75,14 @@ handle_robustness_analysis_command(CmdArgs) ->
             LoadTimeout = timer:seconds(LoadTimeoutSecs),
             LogLevel = list_to_atom(StrLogLevel),
             ok = locus_logger:set_loglevel(LogLevel),
-            prepare_robustness_analysis(DatabaseURL, LoadTimeout);
+            prepare_analysis(DatabaseURL, LoadTimeout);
         _ ->
-            getopt:usage(OptSpecList, "locus analyze-robustness"),
+            getopt:usage(OptSpecList, "locus analyze"),
             fall_from_grace()
     end.
 
-prepare_robustness_analysis(DatabaseURL, LoadTimeout) ->
-    DatabaseId = cli_robustness_check,
+prepare_analysis(DatabaseURL, LoadTimeout) ->
+    DatabaseId = cli_analysis,
     WaitRef = make_ref(),
     BaseOpts = [{async_waiter, {self(),WaitRef}}],
     ExtraOpts =
@@ -94,14 +94,14 @@ prepare_robustness_analysis(DatabaseURL, LoadTimeout) ->
     stderr_println("Loading database from \"~ts\"...", [DatabaseURL]),
     case locus:start_loader(DatabaseId, DatabaseURL, BaseOpts ++ ExtraOpts) of
         ok ->
-            wait_for_robustness_analysis_database_load(DatabaseId, WaitRef, LoadTimeout)
+            wait_for_analysis_database_load(DatabaseId, WaitRef, LoadTimeout)
     end.
 
-wait_for_robustness_analysis_database_load(DatabaseId, WaitRef, LoadTimeout) ->
+wait_for_analysis_database_load(DatabaseId, WaitRef, LoadTimeout) ->
     receive
         {WaitRef, {ok, LoadedVersion}} ->
             stderr_println("Database version ~p successfully loaded", [LoadedVersion]),
-            perform_robustness_analysis(DatabaseId);
+            perform_analysis(DatabaseId);
         {WaitRef, {error, Reason}} ->
             fall_from_grace("Failed to load database: ~p", [Reason])
     after
@@ -109,15 +109,15 @@ wait_for_robustness_analysis_database_load(DatabaseId, WaitRef, LoadTimeout) ->
             fall_from_grace("Timeout loading the database")
     end.
 
-perform_robustness_analysis(DatabaseId) ->
-    stderr_println("Checking database robustness..."),
-    case locus:analyze_robustness(DatabaseId) of
+perform_analysis(DatabaseId) ->
+    stderr_println("Analyzing database for flaws..."),
+    case locus:analyze(DatabaseId) of
         ok ->
             stderr_println("Database is wholesome.");
-        {error, {frail, Frailties}} ->
+        {error, {flawed, Flaws}} ->
             fall_from_grace("Database is corrupt or incompatible:~n"
-                            ++ lists:flatten(["* ~p~n" || _ <- Frailties]),
-                            Frailties)
+                            ++ lists:flatten(["* ~p~n" || _ <- Flaws]),
+                            Flaws)
     end.
 
 fall_from_grace() ->